requirements.txt
discord.py
psycopg2-binary
Pillow
requests
python-dotenv
asyncpg